% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_methods.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Regression coefficients for a linreg object}
\usage{
\method{coef}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class linreg}

\item{...}{Additional arguments to be passed to the print method}
}
\value{
linreg object regression coefficients
}
\description{
Regression coefficients for a linreg object
}
