% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{linreg(formula, data)}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{A linear regression in the form  y ~ var1 + var2 ...}

\item{data}{A data frame on which the regression will be done}
}
\value{
A linreg object
}
\description{
linreg(formula, data)
}
