% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_methods.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Diagnostic plots for a linreg object}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{An object of class linreg}

\item{...}{Additional arguments for compatibility with generic method}
}
\value{
Invisibly a list with two ggplot2 objects
}
\description{
Diagnostic plots for a linreg object
}
