% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_methods.R
\name{pred}
\alias{pred}
\title{Predicted values generic function}
\usage{
pred(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{Additional arguments.}
}
\description{
Predicted values generic function
}
