% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_methods.R
\name{pred.linreg}
\alias{pred.linreg}
\title{Predicted values for a linreg object}
\usage{
pred.linreg(object)
}
\arguments{
\item{object}{An object of class linreg}
}
\value{
A linreg object predicted values
}
\description{
Predicted values for a linreg object
}
