% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_methods.R
\name{pred.linreg}
\alias{pred.linreg}
\title{Predicted values for a linreg object}
\usage{
\method{pred}{linreg}(object)
}
\arguments{
\item{object}{A linreg object}
}
\value{
linreg object predicted values
}
\description{
Predicted values for a linreg object
}
