% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_methods.R
\name{print.linreg}
\alias{print.linreg}
\title{Print for a linreg object}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{An object of class linreg}

\item{...}{Additional arguments to be passed to the print method}
}
\value{
Invisibly the linreg object
}
\description{
Prints the regression coefficients of a linreg object
}
