% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_methods.R
\name{residuals.linreg}
\alias{residuals.linreg}
\title{Residuals for a linreg object}
\usage{
\method{residuals}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class linreg.}

\item{...}{Additional arguments for compatibility with generic method}
}
\value{
An array of residuals as the difference between predicted and actual response values
}
\description{
Residuals for a linreg object
}
