% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_methods.R
\name{summary.linreg}
\alias{summary.linreg}
\title{Summary of the regression results for a linreg object}
\usage{
\method{summary}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class linreg}

\item{...}{Additional arguments for compatibility with generic method}
}
\value{
Invisibly the linreg object
}
\description{
Summary of the regression results for a linreg object
}
